Select reaction (#3)
Add a selection menu to that displays selected reaction

* Add mostly working menu

* Engine class addition and some code cleanup

Added the engine class that takes the information provided by the data classes and turns it into SceneKit nodes that can then be rendered in ViewController. Currently does not animate or distinguish between bond types. That functionality will be added once bond types are added into the data classes.

* Removed junk

* Minor fix

* Support reading partial bonds

* Update gitignore

* Merge from Engine-Testing

* Add reactions, make reaction accessible

* Add selection menu and more reaction files

* Adds temp(?) fix to reset scene

* Fix missing atoms

* Update scene lighting

* Fixed missing buttons

* Fix blank space

+++ MoleculeDemo4/MoleculeDemo3/dataObjects/Bond.swift
@@ -8,9 +8,9 @@
 
 public class Bond {
     public let atom1, atom2: Atom
-    public let order: Int
+    public let order: Double
 
-    init(_ atom1: Atom, _ atom2: Atom, _ order: Int) {
+    init(_ atom1: Atom, _ atom2: Atom, _ order: Double) {
         self.atom1 = atom1
         self.atom2 = atom2
         self.order = order
